code refactors, compile time improvements, math improvements

+++ src/lin-alloc.c
@@ -9,7 +9,7 @@ typedef struct sx__linalloc_hdr_s {
     int padding;    // number of bytes that is padded before the pointer
 } sx__linalloc_hdr;
 
-static void* sx__linalloc_malloc(sx_linalloc* alloc, size_t size, size_t align) {
+static void* sx__linalloc_malloc(sx_linalloc* alloc, size_t size, uint32_t align) {
     align = sx_max((int)align, SX_CONFIG_ALLOCATOR_NATURAL_ALIGNMENT);
     size_t total = size + sizeof(sx__linalloc_hdr) + align;
     if (alloc->offset + total > alloc->size) {
@@ -31,7 +31,7 @@ static void* sx__linalloc_malloc(sx_linalloc* alloc, size_t size, size_t align) 
     return aligned;
 }
 
-static void* sx__linalloc_cb(void* ptr, size_t size, size_t align, const char* file,
+static void* sx__linalloc_cb(void* ptr, size_t size, uint32_t align, const char* file,
                              const char* func, uint32_t line, void* user_data) {
     sx_linalloc* linalloc = (sx_linalloc*)user_data;
     void*        last_ptr = linalloc->ptr + linalloc->last_ptr_offset;
